+++ src/commonMain/kotlin/csense/kotlin/annotations/ExperimentalCsenseApi.kt
@@ -1,0 +1,5 @@
+package csense.kotlin.annotations
+
+
+@RequiresOptIn("Very experimental, due to global namespace pollution", RequiresOptIn.Level.WARNING)
+public annotation class ExperimentalCsenseApi