+++ src/commonMain/kotlin/csense/kotlin/extensions/collections/array/BooleanArray.kt
@@ -16,22 +16,6 @@ public inline fun <U> BooleanArray.forEachDiscard(receiver: Function1<Boolean, U
 
 //region Generic collection extensions
 /**
- * Performs traversal in pairs of 2  (with the first index as well)
- */
-@Deprecated("will be removed")
-@Suppress("MissingTestFunction")
-public inline fun BooleanArray.forEach2Indexed(action: Function2IndexedUnit<Boolean, Boolean>): Unit =
-    GenericCollectionExtensions.forEach2Indexed(count(), ::elementAt, action)
-
-/**
- * Performs traversal in pairs of 2
- */
-@Deprecated("will be removed")
-@Suppress("MissingTestFunction")
-public inline fun BooleanArray.forEach2(action: Function2Unit<Boolean, Boolean>): Unit =
-    GenericCollectionExtensions.forEach2(count(), ::elementAt, action)
-
-/**
  * Performs backwards traversal on this [BooleanArray].
  */
 public inline fun BooleanArray.forEachBackwards(action: FunctionUnit<Boolean>): Unit =