+++ src/commonMain/kotlin/csense/kotlin/extensions/collections/array/IntArray.kt
@@ -16,22 +16,6 @@ public inline fun <U> IntArray.forEachDiscard(receiver: Function1<Int, U>): Unit
 
 //region Generic collection extensions
 /**
- * Performs traversal in pairs of 2  (with the first index as well)
- */
-@Deprecated("will be removed")
-@Suppress("MissingTestFunction")
-public inline fun IntArray.forEach2Indexed(action: Function2IndexedUnit<Int, Int>): Unit =
-    GenericCollectionExtensions.forEach2Indexed(count(), ::elementAt, action)
-
-/**
- * Performs traversal in pairs of 2
- */
-@Deprecated("will be removed")
-@Suppress("MissingTestFunction")
-public inline fun IntArray.forEach2(action: Function2Unit<Int, Int>): Unit =
-    GenericCollectionExtensions.forEach2(count(), ::elementAt, action)
-
-/**
  * Performs backwards traversal on this [IntArray].
  */
 public inline fun IntArray.forEachBackwards(action: FunctionUnit<Int>): Unit =