+++ src/commonMain/kotlin/csense/kotlin/extensions/collections/map/Map.kt
@@ -25,26 +25,6 @@ public inline fun <K, V> Map<K, V>.forEachIndexed(action: (Map.Entry<K, V>, Int)
 
 //region Generic collection extensions
 /**
- * Performs traversal in pairs of 2  (with the first index as well)
- * @receiver [List]<T>
- * @param action [Function2IndexedUnit]<T, T>
- */
-@Deprecated("will be removed")
-@Suppress("MissingTestFunction")
-public inline fun <K, V> Map<K, V>.foreach2Indexed(action: Function2IndexedUnit<Map.Entry<K, V>, Map.Entry<K, V>>): Unit =
-    GenericCollectionExtensions.forEach2Indexed(size, this.entries::elementAt, action)
-
-/**
- * Performs traversal in pairs of 2
- * @receiver [List]<T>
- * @param action [Function2Unit]<T, T>
- */
-@Deprecated("will be removed")
-@Suppress("MissingTestFunction")
-public inline fun <K, V> Map<K, V>.foreach2(action: Function2Unit<Map.Entry<K, V>, Map.Entry<K, V>>): Unit =
-    GenericCollectionExtensions.forEach2(size, this.entries::elementAt, action)
-
-/**
  * Performs backwards traversal on this
  * @receiver [List]<T>
  * @param action [FunctionUnit]<T>