+++ src/commonMain/kotlin/csense/kotlin/extensions/collections/map/MutableMap.kt
@@ -71,3 +71,8 @@ public inline fun <Key, Value> MutableMap<Key, Value>.reverse(): MutableMap<Valu
     }
     return result
 }
+
+
+public inline fun <Key, Value> MutableMap<Key, Value>.remove(key: Key?): Value? {
+    return key?.let(::remove)
+}