+++ src/commonMain/kotlin/csense/kotlin/extensions/collections/set/Set.kt
@@ -7,31 +7,6 @@ import csense.kotlin.extensions.collections.generic.*
 
 
 //region Generic collection extensions
-/**
- * Performs traversal in pairs of 2  (with the first index as well)
- * @receiver [List]<T>
- * @param action [Function2IndexedUnit]<T, T>
- * @timecomplexity O(N)
- */
-@Deprecated("will be removed")
-@Suppress("MissingTestFunction")
-public inline fun <T> Set<T>.foreach2Indexed(action: Function2IndexedUnit<T, T>): Unit =
-    GenericCollectionExtensions.forEach2Indexed(
-        size,
-        ::elementAt,
-        action
-    )
-
-/**
- * Performs traversal in pairs of 2
- * @receiver [List]<T>
- * @param action [Function2Unit]<T, T>
- * @timecomplexity O(N)
- */
-@Deprecated("will be removed")
-@Suppress("MissingTestFunction")
-public inline fun <T> Set<T>.foreach2(action: Function2Unit<T, T>): Unit =
-    GenericCollectionExtensions.forEach2(size, ::elementAt, action)
 
 /**
  * Performs backwards traversal on this