+++ src/commonTest/kotlin/csense/kotlin/extensions/collections/map/MutableMapKtTest.kt
@@ -238,4 +238,87 @@ class MutableMapKtTest {
         }
 
     }
+
+    class MutableMapKeyValueRemove {
+        class MutableMapKeyValueRemove {
+
+            @Test
+            fun emptyNull() {
+                mutableMapOf<String, String>().apply {
+                    remove(null).assertNull()
+                    assertSize(0)
+                }
+            }
+
+            @Test
+            fun emptyValue() {
+                mutableMapOf<String, String>().apply {
+                    remove("abc").assertNull()
+                    assertSize(0)
+                }
+            }
+
+
+            @Test
+            fun singleNull() {
+                mutableMapOf("a" to "1").apply {
+                    remove(null).assertNull()
+                    assertSize(1)
+                    assertContains("a" to "1")
+                }
+            }
+
+            @Test
+            fun singleNotFound() {
+                mutableMapOf("a" to "1").apply {
+                    remove("abc").assertNull()
+                    assertSize(1)
+                    assertContains("a" to "1")
+                }
+            }
+
+
+            @Test
+            fun singleFound() {
+                mutableMapOf("a" to "1").apply {
+                    remove("a").assertNotNullAndEquals("1")
+                    assertEmpty()
+                }
+            }
+
+
+            @Test
+            fun multipleNull() {
+                mutableMapOf("a" to "1", "b" to "2").apply {
+                    remove(null).assertNull()
+                    assertSize(2)
+                    assertContains("a" to "1")
+                    assertContains("b" to "2")
+                }
+            }
+
+            @Test
+            fun multipleNotFound() {
+                mutableMapOf("a" to "1", "b" to "2").apply {
+                    remove("q").assertNull()
+                    assertSize(2)
+                    assertContains("a" to "1")
+                    assertContains("b" to "2")
+                }
+            }
+
+            @Test
+            fun multipleFound() {
+                mutableMapOf("a" to "1", "b" to "2").apply {
+                    remove("b").assertNotNullAndEquals("2")
+                    assertSingle {
+                        it.key.assert("a")
+                        it.value.assert("1")
+                    }
+                }
+            }
+
+        }
+
+    }
 }