+++ src/commonTest/kotlin/csense/kotlin/extensions/coroutines/CoroutineScopeTest.kt
@@ -1,16 +1,17 @@
+@file:OptIn(ExperimentalCoroutinesApi::class)
+
 package csense.kotlin.extensions.coroutines
 
-import csense.kotlin.coroutines.*
 import csense.kotlin.tests.assertions.*
 import kotlinx.coroutines.*
+import kotlinx.coroutines.test.*
+import kotlin.jvm.*
 import kotlin.test.*
 
-//see https://github.com/Kotlin/kotlinx.coroutines/issues/1996 for Main issue(s)
 class CoroutineScopeTest {
 
-
     @Test
-    fun coroutineScopeAsyncDefault() = runBlockingTest {
+    fun coroutineScopeAsyncDefault() = runTest {
         val async = asyncDefault {
             assertDispatcher(Dispatchers.Default)
             "result"
@@ -20,7 +21,7 @@ class CoroutineScopeTest {
 
 
     @Test
-    fun coroutineScopeWithContextDefault() = runBlockingTest {
+    fun coroutineScopeWithContextDefault() = runTest {
         withContextDefault {
             assertDispatcher(Dispatchers.Default)
             "result"
@@ -29,7 +30,7 @@ class CoroutineScopeTest {
 
 
     @Test
-    fun coroutineScopeLaunchDefault() = runBlockingTest {
+    fun coroutineScopeLaunchDefault() = runTest {
         assertCalled { shouldBeCalled ->
             launchDefault {
                 assertDispatcher(Dispatchers.Default)
@@ -37,20 +38,35 @@ class CoroutineScopeTest {
             }.join()
         }
     }
-
-    @Test
-    fun coroutineScopeWithContextMain() {
-        @Suppress("UNUSED_VARIABLE") val x = 0 //Due to platform dependent tests
-    }
-
-    @Test
-    fun coroutineScopeLaunchMain() = runBlockingTest {
-        @Suppress("UNUSED_VARIABLE") val x = 0 //Due to platform dependent tests
-    }
-
-    @Test
-    fun coroutineScopeAsyncMain() = runBlockingTest {
-        @Suppress("UNUSED_VARIABLE") val x = 0 //Due to platform dependent tests
-    }
+//
+//    @Test
+//    fun coroutineScopeWithContextMain() = runTest {
+//        withContextMain {
+//            assertDispatcher(Dispatchers.Main)
+//            "result"
+//        }.assert("result")
+//        mainThreadSurrogate.scheduler.runCurrent()
+//    }
+//
+//    @Test
+//    fun coroutineScopeLaunchMain() = runTest {
+//        assertCalled { shouldBeCalled ->
+//            launchMain {
+//                assertDispatcher(Dispatchers.Main)
+//                shouldBeCalled()
+//            }.join()
+//        }
+//        mainThreadSurrogate.scheduler.runCurrent()
+//    }
+//
+//    @Test
+//    fun coroutineScopeAsyncMain() = runTest {
+//        val async = asyncMain {
+//            assertDispatcher(Dispatchers.Main)
+//            "result"
+//        }
+//        async.await().assert("result")
+//        mainThreadSurrogate.scheduler.runCurrent()
+//    }
 
 }