+++ changelog.md
@@ -3,7 +3,23 @@
 ## 0.0.54
 
 - added
-    - Map & MutableMap reverse (reverses the relation, so from key -> value it is value -> key)
+    - Map
+        - reverse (reverses the relation, so from key -> value it is value -> key)
+      
+    - MutableMap
+        - reverse (reverses the relation, so from key -> value it is value -> key)
+        - remove (key?) convince remove function
+      
+    - @ExperimentalCsenseApi for marking things that are really experimental and might be removed or changed without
+      notice
+
+- Removed
+    - deprecated things (most had a deprecation cycle of 4+ releases)
+        - if using the foreach2 extensions the general function is still available at GenericCollectionExtensions.{name}
+  
+    - Workaround for coroutines tests (runTest)
+
+    
 
 ## 0.0.53